fix: allow decimal protein values
FoodItemBase and FoodItemUpdate typed protein as int, so non-whole
values (e.g. 37.5g from a 1.5-scoop protein shake quick-log) failed
validation on create/update.

+++ app.py
@@ -161,7 +161,7 @@ def delete_food(
 class FoodItemUpdate(BaseModel):
     name: Optional[str] = None
     calories: Optional[int] = None
-    protein: Optional[int] = None
+    protein: Optional[float] = None
     carbs: Optional[int] = None
     fat: Optional[int] = None
     category: Optional[str] = None
